+++ app/templates/base.html
@@ -55,11 +55,9 @@
 
     }
     #video {
-        height: 80%;
         margin-bottom: 5px;
     }
     #videoControl {
-        height: 20%;
         -moz-border-radius-topleft: 50px;
         border-top-left-radius: 50px;
         vertical-align: center;
@@ -73,9 +71,12 @@
         text-align: center;
     }
 .videosrc {
+    width: 100%;
     max-height: 100%;
     -moz-border-radius-bottomleft: 50px;
     border-bottom-left-radius: 50px;
+    -moz-border-radius-topleft: 50px;
+    border-top-left-radius: 50px;
 
 }
 </style>
@@ -83,7 +84,7 @@
 <body>
     <style>
         body{
-          background-image: url('static/webasset/img/see_bg.png');
+         background-image: url('static/webasset/img/seeBGalternative.jpg');
         }
     </style>
     <table class="main-layout">
