+++ app/templates/frameUnit/__settings.html
@@ -1,0 +1,3 @@
+<!-- Settings template to return via AJAX call -->
+<h1>My settings</h1>
+        <button onclick="home()">Go home</button>
