+++ app/templates/login.html
@@ -62,7 +62,7 @@
     left: 0;
     background-size: cover;
     background-position: 50% 50%;
-    background-image: url('static/webasset/img/see_bg.png');
+    background-image: url('static/webasset/img/seeBGalternative.jpg');
     background-repeat:repeat;
   }
   .form-signin {
